refactor(lox-frames): merge the Frame enum into the frames module
